requirements.txt
streamlit
fastapi
uvicorn
requests
sentence-transformers
faiss-cpu
pypdf
easyocr
pdf2image
